Don't parse word time tags if multiple line times are present
Word time tags are absolute timestamps in the lyrics, so having multiple line times would generally conflict with them. Thus, they should be just ignored instead, since the format is wrong.
The same applies if there is no line time tag, in that case, it wouldn't be known when the first word starts.
Both are inconsidered incorrect LRC and thus only parsed on a best-effort basis.